Merge tuples of transitively similar images in dictionary
The current implementation of -[OSSimilaritySearch dictionaryFromSimilarImagesResult:] does not handle tuples of transitively similar images. Consider three images A, B, and C similar to each other. Similarity search will produce the tuples (A, B), (A, C), and (B, C), which are then transformed into the dictionary representation {A: [B, C], B: [C]}, whereas one would expect a representation like {A: [B, C]}.

To handle transitivity, this patch introduces a simplified union-find-like approach (another way being, for example, traversing connected components in a graph-like structure). The optimization is possible since the similarImages* methods already return the list of tuples in a strictly ordered permutation, e. g. for the images above there cannot be a tuple (C, A).